requirements.txt
Flask==2.2.5
flask-cors==3.0.10
pymongo==4.3.3
requests==2.28.1
Pillow==9.4.0
python-dotenv==0.21.0
gunicorn==20.1.0
imaginairy==9.6.0